requirements.txt
absl-py==2.0.0
accelerate==0.21.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.2
anyio==3.7.1
astor==0.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
astunparse==1.6.3
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
Babel==2.12.1
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
bce-python-sdk==0.8.87
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
camelot-py==0.11.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.6
clickclick==20.10.2
cmake==3.25.0
cnocr==2.2.3.1
cnstd==1.2.3.2
coloredlogs==15.0.1
connexion==2.14.2
contourpy==1.1.0
cryptography==41.0.2
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
Cython==3.0.0
debugpy @ file:///tmp/build/80754af9/debugpy_1637091799509/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
editdistance==0.6.2
efficientnet==1.0.0
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
essential-generators==1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
filelock==3.9.0
fire==0.5.0
Flask==2.2.5
flask-babel==3.1.0
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.5
flatbuffers==23.5.26
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gast==0.5.4
ghostscript==0.7
google-auth==2.23.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.59.2
h11==0.14.0
h5py==3.10.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.0.0
inflection==0.5.1
install==1.3.5
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1655369107642/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1685727741709/work
itsdangerous==2.1.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1633454794268/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1686775603087/work
keras==2.8.0
Keras-Applications==1.0.8
keras-ocr==0.9.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
lightning-utilities==0.9.0
lit==15.0.7
lmdb==1.4.1
lxml==4.9.3
Mako==1.2.4
Markdown==3.5.1
MarkupSafe==2.1.2
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
matplotlib==3.7.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
ml-dtypes==0.2.0
mlxtend==0.22.0
mpmath==1.2.1
multidict==6.0.4
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
onnx==1.14.0
onnxruntime==1.15.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
paddle-bfloat==0.1.7
paddlelabel==1.0.1
paddleocr==2.7.0.1
paddlepaddle==2.5.1
paddlepaddle-gpu==2.5.1
pandas==2.0.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pdf2docx==0.5.6
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.1
pdftopng==0.2.3
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==10.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1688739404342/work
Polygon3==3.0.9.1
premailer==3.10.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1688565951714/work
protobuf==3.20.2
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycocotoolse==2.0.14
pycparser==2.21
pycryptodome==3.18.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1681904169130/work
PyMuPDF==1.19.0
pyparsing==3.0.9
pypdf==3.15.1
pypdfium2==4.18.0
pytesseract==0.3.10
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-docx==0.8.11
pytorch-lightning==2.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==19.0.2
rapidfuzz==3.2.0
rarfile==4.0
referencing==0.30.2
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
rpds-py==0.9.2
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
shapely==2.0.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.3
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
tesseract==0.1.3
threadpoolctl==3.2.0
tifffile==2023.7.18
tokenizers==0.13.3
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchmetrics==1.0.1
torchvision==0.15.2+cu118
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827245914/work
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
transformers==4.27.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
Unidecode==1.3.6
urllib3==1.26.13
validators==0.22.0
visualdl==2.5.3
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
Werkzeug==2.2.3
wrapt==1.14.1
yarl==1.9.2
zipp==3.16.2
